+++ frontend/src/index.css
@@ -14,9 +14,6 @@
   font-weight: 400;
 
   color-scheme: light dark;
-  color: #213547;
-  background-color: #ffffff;
-
   font-synthesis: none;
   text-rendering: optimizeLegibility;
   -webkit-font-smoothing: antialiased;
@@ -42,6 +39,9 @@ h1 {
 }
 
 #root {
+  width: 100%;
+  height: 100%;
   min-width: 100vw;
   min-height: 100vh;
+  background-color: #263238;
 }
